requirements.txt
aiohttp==3.6.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
discord.py==1.5.1
idna==2.10
isort==5.6.4
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.3.2
mccabe==0.6.1
multidict==4.7.6
numpy==1.19.4
onnx==1.7.0
onnxruntime==1.5.2
Pillow==8.0.1
protobuf==3.13.0
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
wrapt==1.12.1
yarl==1.5.1
